Nathan add logging to metrics (#157)
what this PR does:

If you want to log out something comming from the metrics, simply return it in the metric dict.
for example, if you want to log out the judge response when using llm_as_judge, simply return the response in the dict.

```
{
   "score": score,
   "judgement": judge_response
}
````

the `judgement` field is a string and will not be aggregated. however, it will be logged in the details for each sample.

---------

Co-authored-by: Nathan Habib <nathan.habib@huggingface.com>